implement dataset/detection.py main to compute dataset's stats
Implemented main in order to retrieve the statistics of the dataset and changed of detector_metrics class, computing iou and map at the same time

+++ src/utils.py
@@ -83,10 +83,10 @@ def handle_dataset_kaggle(images_dir, annotations_dir):
 
             # Get bounding box coordinates
             bndbox = obj.find("bndbox")
-            xmin = int(bndbox.find("xmin").text)
-            ymin = int(bndbox.find("ymin").text)
-            xmax = int(bndbox.find("xmax").text)
-            ymax = int(bndbox.find("ymax").text)
+            xmin = float(bndbox.find("xmin").text)
+            ymin = float(bndbox.find("ymin").text)
+            xmax = float(bndbox.find("xmax").text)
+            ymax = float(bndbox.find("ymax").text)
 
             # Append bounding box and class to the list
             bboxes.append({
